+++ pkg/security/auth0/verifier_test.go
@@ -8,7 +8,6 @@ import (
 
 	"github.com/alexfalkowski/go-service/pkg/cache/ristretto"
 	"github.com/alexfalkowski/go-service/pkg/logger/zap"
-	"github.com/alexfalkowski/go-service/pkg/meta"
 	"github.com/alexfalkowski/go-service/pkg/security/auth0"
 	"github.com/alexfalkowski/go-service/pkg/transport/http"
 	. "github.com/smartystreets/goconvey/convey"
@@ -41,9 +40,9 @@ func TestVerify(t *testing.T) {
 		lc.RequireStart()
 
 		Convey("When I verify the token", func() {
-			ctx := meta.WithAttribute(context.Background(), meta.RequestID, "test-request-id")
+			ctx := context.Background()
 
-			token, err := gen.Generate(ctx)
+			token, err := gen.Generate(context.Background())
 			So(err, ShouldBeNil)
 
 			err = ver.Verify(ctx, token)
